requirements.txt
asgiref==3.10.0
certifi==2025.11.12
charset-normalizer==3.4.4
Django==5.2.7
django-debug-toolbar==6.1.0
gunicorn==23.0.0
idna==3.11
packaging==25.0
pillow==12.0.0
PyJWT==2.10.1
requests==2.32.5
ruff==0.14.5
sqlparse==0.5.3
urllib3==2.6.2
redis==5.3.1
